:wrench:[Chore] work 생성하기 url 루트 변경
