+++ src/pages/work/[id]/index.jsx
@@ -15,8 +15,8 @@ export default function WorkDetailPage() {
   if (isPending) {
     return <Loader />;
   }
-  // console.log(data);
-  // const { isClosed } = data;
+  console.log(data);
+  const { isClosed } = data;
   return (
     <>
       <Head>
